pyproject.toml
[tool.poetry]
name = "hilopalace"
version = "0.1.0"
description = "Scrape and reformat a non-responsive website"
authors = ["Allan Anderson <rubel@goosemoon.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-tailwind = "^3.5.0"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
